+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added quick mode to spaceload report in myems-api.
 - Added object ID to editor title of model dialogs in myems-admin.
 - Added ID column to MeterBatch report in myems-web.
+- Added ID column to MeterTracking report in myems-api and myems-web.
 
 ### Changed
 - Updated translations of myems-admin.